pyproject.toml
[project]
name = "yapper"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.8.0",
    "apscheduler==3.11.0",
    "attrs==25.1.0",
    "blinker==1.9.0",
    "certifi==2025.1.31",
    "cffi==1.17.1",
    "charset-normalizer==3.4.1",
    "chroma-hnswlib==0.7.6",
    "click==8.1.8",
    "colorama==0.4.6",
    "cryptography==44.0.2",
    "distro==1.9.0",
    "durationpy==0.9",
    "filelock==3.17.0",
    "Flask>=2.3.0,<3.0",
    "flask-apscheduler==1.13.1",
    "flask-cors==3.0.10",
    "flatbuffers==25.2.10",
    "fsspec==2025.2.0",
    "greenlet==3.1.1",
    "gunicorn==23.0.0",
    "h11==0.14.0",
    "httpcore==1.0.7",
    "httpx>=0.26.0",
    "idna==3.10",
    "iniconfig==2.0.0",
    "itsdangerous==2.2.0",
    "jinja2==3.1.5",
    "jiter==0.8.2",
    "llvmlite==0.44.0",
    "lxml==5.3.1",
    "mako==1.3.9",
    "markupsafe==3.0.2",
    "monotonic==1.6",
    "more-itertools==10.6.0",
    "mpmath==1.3.0",
    "mutagen>=1.47.0",
    "networkx==3.4.2",
    "numba==0.61.0",
    "numpy==2.1.3",
    "openai==1.64.0",
    "openai-whisper>=20240930",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "opentelemetry-util-http==0.51b0",
    "orjson==3.10.15",
    "overrides==7.7.0",
    "packaging==24.2",
    "pgvector==0.3.6",
    "pip==25.0.1",
    "pluggy==1.5.0",
    "protobuf==5.29.3",
    "pyasn1==0.6.1",
    "pycparser==2.22",
    "pydantic>=2.0.0",
    "pydantic-core==2.27.2",
    "pygments==2.19.1",
    "pyjwt",
    "pypika==0.48.9",
    "pyproject-hooks==1.2.0",
    "pyreadline3==3.5.4",
    "pytest==7.2.2",
    "pytest-cov>=7.0.0",
    "python-dateutil==2.9.0.post0",
    "python-docx==1.1.2",
    "python-dotenv==1.0.0",
    "python-multipart>=0.0.6",
    "pytz==2025.1",
    "pyyaml==6.0.2",
    "regex==2024.11.6",
    "requests",
    "setuptools==75.8.2",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "supabase>=2.16.0",
    "sympy==1.13.1",
    "tenacity==9.0.0",
    "tiktoken==0.9.0",
    "torch==2.2.0",
    "tqdm==4.67.1",
    "typing-extensions>=4.8.0",
    "tzdata==2025.1",
    "tzlocal==5.3",
    "urllib3==2.3.0",
    "websocket-client==1.8.0",
    "websockets>=12.0",
    "Werkzeug>=2.3.0,<3.0",
    "wheel==0.45.1",
    "wrapt==1.17.2",
    "zipp==3.21.0",
]
